fix(screen): make the screens begin to work again

+++ src/main/java/com/mcf/davidee/nbtedit/gui/GuiNBTButton.java
@@ -59,7 +59,7 @@ public class GuiNBTButton extends Widget {
 		return enabled;
 	}
 
-	public boolean inBounds(int mx, int my) {
+	public boolean inBounds(double mx, double my) {
 		return enabled && mx >= x && my >= y && mx < x + WIDTH && my < y + HEIGHT;
 	}
